requirements.txt
numpy
torch
torchvision
streamlit
pillow
opencv-python